refactor(plugins): extract tool visibility collector below file size limit
